backend/requirements.txt
# Python FastAPI Backend Requirements
fastapi==0.95.1
uvicorn[standard]==0.22.0
pydantic==1.10.8
requests==2.31.0
numpy==1.26.0
pandas==2.1.1
sqlalchemy==2.0.21
bcrypt
pyjwt

# Testing 
httpx
pytest
pytest-dotenv